# api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: '0.0.1'
  title: Luther API
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: |
      For accessing the API a valid JWT token must be passed in all the queries in
      the 'Authorization' header.


      A valid JWT token is generated by the API and retourned as answer of a call
      to the route /signin giving a valid user & password.


      The following syntax must be used in the 'Authorization' header :

        xxxxxx.yyyyyyy.zzzzzz
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /user:
    x-swagger-router-controller: user
    get:
      description: Get User By Email
      tags:
        - user
      operationId: getUser
      parameters:
        - name: email
          in: query
          description: user's email
          required: true
          type: string
      security:
        - Bearer: []
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/UserResponse'
    put:
      description: Update User
      tags:
        - user
      operationId: updateUser
      parameters:
        - name: id
          in: query
          description: user's id
          required: true
          type: string
        - name: body
          in: body
          required: true
          schema:
            properties:
              first_name:
                type: string
              last_name:
                type: string
              email:
                type: string
              password:
                type: string
              roles:
                type: array
                items:
                  type: string
      security:
        - Bearer: []
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/UserResponse'
        '409':
          description: 'Email Already Registered'
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
    post:
      description: Add New User
      tags:
        - user
      operationId: signup
      parameters:
        - name: body
          in: body
          required: true
          schema:
            properties:
              first_name:
                type: string
              last_name:
                type: string
              email:
                type: string
              password:
                type: string
              roles:
                type: array
                items:
                  type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/UserResponse'
        '409':
          description: 'Email Already Registered'
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: 'No exist'
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
    delete:
      description: Delete User
      tags:
        - user
      operationId: deleteUser
      parameters:
        - name: id
          in: query
          description: user's id
          required: true
          type: string
      security:
        - Bearer: []
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/UserResponse'
        '404':
          description: 'No exist'
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
  /user/activate:
    x-swagger-router-controller: user
    put:
      description: Activate User
      tags:
        - user
      operationId: activate
      parameters:
        - name: id
          in: query
          description: user's id
          required: true
          type: string
      security:
        - Bearer: []
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/UserResponse'
        '409':
          description: 'Invalid id'
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
  /user/deactivate:
    x-swagger-router-controller: user
    put:
      description: Deactivate User
      tags:
        - user
      operationId: deactivate
      parameters:
        - name: id
          in: query
          description: user's id
          required: true
          type: string
      security:
        - Bearer: []
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/UserResponse'
        '409':
          description: 'Invalid id'
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
  /user/invite:
    x-swagger-router-controller: user
    post:
      description: Invite User
      tags:
        - user
      operationId: inviteUser
      parameters:
        - name: body
          in: body
          required: true
          schema:
            properties:
              first_name:
                type: string
              last_name:
                type: string
              email:
                type: string
              password:
                type: string
              roles:
                type: array
                items:
                  type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/UserResponse'
        '409':
          description: 'Email Already Registered'
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: 'No exist'
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
...
  /user/list:
    x-swagger-router-controller: user
    get:
      description: List users
      tags:
        - user
      operationId: listUsers
      security:
        - Bearer: []
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/UsersResponse'
...
definitions:
  UserResponse:
    type: object
    properties:
      _id:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      email:
        type: string
      password:
        type: string
      roles:
        type: array
        items:
          type: string
      date_created:
        type: string
      is_deleted:
        type: boolean
  UsersResponse:
    type: array
    items:
      type: object
      properties:
        _id:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        password:
          type: string
        roles:
          type: array
          items:
            type: string
        date_created:
          type: string
        is_deleted:
          type: boolean
...
  Error:
    type: object
    properties:
      message:
        type: string
#TAGS
...